rename dial package to kademliaclient
